requirements.txt
fastapi==0.109.2
starlette==0.36.3
pydantic==2.5.3
uvicorn==0.27.1
typing-extensions==4.9.0
google-cloud-aiplatform
google-cloud-bigquery
python-docx
google-cloud-storage
